# api/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/core/v1/entity.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/schemas": {
      "post": {
        "summary": "Create a schema",
        "operationId": "CreateSchema",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1CreateSchemaResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "schema value",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "description": "schema value"
            }
          }
        ],
        "tags": [
          "Schema"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateSchemaResponse": {
      "type": "object"
    },
...
  }
}